Third Commit - Header and Footer Includes
This commit enhances the previously created .jsp template by including a
bootstrap powered header that collapses depending on screen-size, and a
footer that should remain on the bottom of the content of the page.

- Created main-header-include.jsp
- main-header-include.jsp has a placeholder menu and search bar that can
be removed/ changed as needed. Header is collapsable for smaller
screens. No functionality in links or search bar.
- Created main-footer-include.jsp
- main-footer-include.jsp has a div with placeholder text
- Removed and replaced some previously created HTML/ JSP in
master-template.jsp. Added filler text as a test, mainly to ensure
footer works correctly.